Support for arm32v7 (#25)
* Move Dockerfile.patch files to architecture dependent directories

Need a way to distinguish between patches when we implement support for other architectures.
Putting them in different paths seems simple and straight forward enough.

* Add -a option to support arm32v7 and amd64 architectures

Previous releases lacked support for compiling for arm32v7 architecture.
In order to keep previous behaviour the default is set to amd64. Using the
-a option, users can now choose between arm32v7 and amd64. If arm32v7 is
used only the Buster release of debian is supported. The limitation is
due to the libmariadb3:armhf dependency not being avaliable in earlier
debian releases.

* Update documentation to detail the -a option

* Remove reference to local dirs

* Set debian package to report correct architecture for arm32v7 build

* Format arm32v7 Dockerfile.patch to match the amd64

* Handle architecture property in control file dynamically

Similar to version, the architecture property in the control file should be
handled dynamically. The build script will now replace the architecture prop
from control.dist to match the choosen architecture.