Use sysusers and tmpfiles to create users and data directories
Instead of manually creating the users, groups, and data directories for vaultwarden during postinst instead use sysusers and tmpfiles. This is more reliable in the long run and better integrated with Systemd.

Additionally fix install logic so that daemon-reload happens automatically and so the vaultwarden service is restarted automatically if it was previously running. The service file will additionally be properly cleaned up during package removal.