Build package from release docker image
This is a radical change in the approach, which come with several
improvements. There is no patching of upstream Dockerfiles anymore,
this greatly reduces coupling. Also, there is no local build, which
leads to much faster package building.

I believe this is the way forward if we don't want to spend too much
time maintaining this.
It comes at a small cost: there is no building under the previous name.